gen_data/requirements.txt
Faker==24.4.0
load-dotenv==0.1.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
